MyFitness |:heavy_check_mark:| Corrigido erros nos cálculos de calorias baseado no sexo do individuo.

+++ database/migrations/2022_11_23_075504_create_basal_metabolic_rates_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('user_name');
             $table->string('gender');
             $table->integer('age');
-            $table->integer('weight');
-            $table->integer('stature');
+            $table->decimal('weight');
+            $table->decimal('stature');
             $table->decimal('activity_rate_factor');
             $table->string('objective');
             $table->string('type_of_diet');
